Add files via upload
STATEMENT:- We have often observed that people tend to waste a lot of time in searching for a suitable area for parking their respective vehicles. At almost every parking area be it in our apartments or shopping complexes or other public parking areas, we see at least one person appointed to guiding the drivers of the incoming vehicles for parking. Through this project I have suggested a cure to these recent day problems and a better management of the parking areas.